Refine equality implementation

+++ dust-lang/tests/operations.rs
@@ -5,7 +5,7 @@ fn if_expression() {
     assert_eq!(run("if true { 1 }"), Ok(Some(Value::integer(1))));
     assert_eq!(
         run("if 42 == 42 { 1 } else { 2 }"),
-        Ok(Some(Value::integer(2)))
+        Ok(Some(Value::integer(1)))
     );
 }
